[BACKLOG-43319] - Name value null check

+++ core/src/main/java/org/pentaho/platform/dataaccess/datasource/wizard/service/impl/ConnectionService.java
@@ -343,6 +343,10 @@ public class ConnectionService {
   @Path( "/deletebyname" )
   public Response deleteConnectionByName( @QueryParam( "name" ) String name ) throws ConnectionServiceException {
     try {
+      if ( StringUtils.isBlank( name ) ) {
+        throw new ConnectionServiceException( com.google.gwt.http.client.Response.SC_BAD_REQUEST, Messages.getErrorString(
+                "ConnectionServiceImpl.ERROR_0003_UNABLE_TO_GET_CONNECTION", String.valueOf( name ) ) ); //$NON-NLS-1$
+      }
       boolean success = connectionService.deleteConnection( URLDecoder.decode( name, StandardCharsets.UTF_8 ) );
       if ( success ) {
         return Response.ok().build();
@@ -433,6 +437,10 @@ public class ConnectionService {
   public IDatabaseConnection getConnectionByName( @QueryParam( "name" ) String name,
                                                   @DefaultValue( "false" ) @QueryParam( "mask" ) Boolean mask )
     throws ConnectionServiceException {
+    if ( StringUtils.isBlank( name ) ) {
+      throw new ConnectionServiceException( com.google.gwt.http.client.Response.SC_BAD_REQUEST, Messages.getErrorString(
+              "ConnectionServiceImpl.ERROR_0003_UNABLE_TO_GET_CONNECTION", String.valueOf( name ) ) ); //$NON-NLS-1$
+    }
     IDatabaseConnection conn = connectionService.getConnectionByName( URLDecoder.decode( name, StandardCharsets.UTF_8 ) );
     if ( mask ) {
       encryptPassword( conn );
